Add an extension module that wraps copyfile(3)
On macOS, copyfile(3) is an API for copying the contents and metadata
of files and directories. It correctly handles special file flags, ACLs,
resource forks, etc.

The extension module has the goal of providing the same interface as
shutil's family of copy functions:

  - shutil.copy
  - shutil.copy2
  - shutil.copyfile (in this commit)
  - shutil.copyfileobj
  - shutil.copymode
  - shutil.copystat
  - shutil.copytree
  - shutil.move

The module also re-uses exceptions provided by shutil in order to be
backwards compatible with existing code.